Fix The Issue That Occurred After Renaming Library Project
Renaming the Hermes.Library project to Hermes caused errors when looking for references to the old project name.

Since the class library project was still empty, I found the quickest solution to the problem was to delete the project and start it again.

Then I renamed the folder for the whole solution and the solution file.
I hate duplication. The idea that the path to the file is C:\R\Hermes\Hermes\SomeClass.cs gives me goosebumps.
After synchronizing the data, I rename the entire repository so that the name corresponds to the name of the repository on disk.