CLASS LIBRARY IMPLENTATION
=== List of changes in the Hermes class library project ===
* I removed the unnecessary class class1.cs
* I've added an ApiHelper class that implements the IApiHelper interface from the Hermes.Abstractions namespace

=== List of changes in the Hermes.WinFormsUI project ===
* Adding an instance of IApiHelper to the Dashboard.cs partial class

=== Detailed description of changes ===
In the class library project, I first removed the class class1.cs and then created the folder / namespace Abstractions. In the folder just created, I created an IApiHelper interface that defines these methods:
1) public bool IsValidUrl(string url)
2) public Task<string> CallApiAsync(string url, bool needFormattedOutput = true)
The interface is implemented in the public ApiHelper class.

The instance of the ApiHelper class is used in the Dashboard.cs partial class as a private field, and I replaced the lines that I commented with the TODO token by calling the methods of this instance.